backend/requirements-prod.txt
# 生产环境依赖
# 基础依赖
-r requirements.txt

# 生产环境专用
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9

# 监控和日志
prometheus-client==0.19.0
django-prometheus==2.3.1
sentry-sdk[django]==1.40.0

# 性能优化
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# 安全
django-cors-headers==4.3.1
django-csp==3.7

# 系统监控
psutil==5.9.6

# 文件存储 (可选)
boto3==1.34.0
django-storages==1.14.2